Tidied implementation; improved existing and added additional test.
