+++ src/reflect/scala/reflect/internal/Types.scala
@@ -1892,11 +1892,7 @@ trait Types
     }
     override def forceDirectSuperclasses: Unit =
       sym0.rawInfo.decls.foreach { decl =>
-        decl match {
-          case _: ModuleSymbol => decl.rawInfo.forceDirectSuperclasses
-          case _: TermSymbol =>
-          case _ => decl.rawInfo.forceDirectSuperclasses
-        }
+        if(decl.isModule || !decl.isTerm) decl.rawInfo.forceDirectSuperclasses
       }
     override protected def finishPrefix(rest: String) = objectPrefix + rest
     override def directObjectString = super.safeToString
