+++ test/files/run/t10454-1/Test_2.scala
@@ -1,0 +1,9 @@
+object Test extends App {
+  def callMacro[T: TC]: List[String] = implicitly[TC[T]].subclasses
+  assert(callMacro[ADT] == List("class ADTConstructor"))
+}
+
+sealed trait ADT
+object ADT {
+  final case class ADTConstructor() extends ADT
+}
