--- test/files/run/t10454/Test_2.scala
@@ -1,9 +1,0 @@
-object Test extends App {
-  def callMacro[T: TC]: Unit = ()
-  callMacro[ADT]
-}
-
-sealed trait ADT
-object ADT {
-  final case class ADTConstructor() extends ADT
-}
